Tests/LibWeb/Text/expected/css/font-rules-within-grouping-rules.txt

Nested face count: 2
NestedMatchingMediaFont: unloaded
NestedSupportsFont: unloaded
Nested face count after removal: 0
